general: add another zoom level
In icon view, add a smaller zoom level to be able for dense views,
and increase the default padding to allow the labels enough space.

Now levels are 48px, 64px, 96px and 128px for icon view, instead of
only 64px, 96px and 128px, but with the increased padding the 64px and
48px are useful.

List view also gains a bigger level, and they become 16px, 32px, 48px,
64px.

Also, adjust the label max width to be larger, but inside the icon
itself. This fixes the label not taking advantage of all the width the
icon provides, and also a few cases where icons were misaligned.