Provides/Conflicts/Replaces eos-file-manager
To make upgrades (hopefully) possible, declare that the various nautilus
packages Conflict and Replace their corresponding eos-file-manager
packages. That should allow apt and dpkg to order the installation
correctly.

The Provides are added to allow packages that depend on the
eos-file-manager packages (e.g., file-roller on
libeos-file-manager-extension1a) to not be removed on the upgrade.