Reapply upstream translations on top of those we had downstream
With a round-trip through Transifex.

This overwrites some of the translations we had in Transifex
with translations that have since been added/modified upstream.
In this case I suspect that our downstream translations
were just for those that were previously missing upstream,
so I'm going ahead with the upstream translations for now.