Fix plural form translations
Since Transifex enforces the rules on the plural forms,
and thus overwrote the rules in is.po and ro.po,
the plural forms as specified were no longer valid
as the "%d" specifier is required for any plural form
that can represent more than one value.

Though perhaps not the perfect fix, I copied the existing
plural form with "%d" over the invalid entries.

https://phabricator.endlessm.com/T16610