Install Nautilus as the default MIME handler for inode/directory
Normally there are no default handlers explicitly set for this
MIME type, which means that GIO and others (e.g. xdg-utils) will
try to determine a fallback by inspecting the cached list of MIME
handlers available, until finding one that handles that type.

In the world of flatpak /var/lib/flatpak/exports/share is added
to the XDG_DATA_DIRS variable with more priority than /usr/share,
meaning that it's relatively easy to reach a situation where an
application would automatically became the "fallback default" over
one installed in the system for a given MIME type, unless the user
has explicitly defined a default application for it.

This has happened so far with the flatpak for Visual Studio, which
installs itself as a handler for inode/directory, also handled by
Nautilus (which is not a flatpak, and therefore its association to
this MIME type lives in /usr/share/applications/mimeinfo.cache). In
this scenario, installing Visual Studio means that, suddenly, trying
to open a directory using the "default" handler would result in VS
performing the action instead of Nautilus, which is very confusing.

This said, it probably makes sense for VS to install itself as a
handler for this MIME type since it serves a clear purpose [1], and
it also probably makes sense that flatpak installs it exports/share
directory in the XDG_DATA_DIRS with precedence over /usr/share (after
all, in a modern flatpak world you probably want to give priority to
whatever app the user installs explicitly over those in the base OS).

However, for the particular case of inode/directory MIME type, I think
it also makes sense to be able to explicitly "protect" the association
with Nautilus (at least in Endless OS, where it's the main file manager)
so that it's not that easy to replace it with another application just
because of being installed, without explicit user intervention.

For this reason, let's ensure that Nautilus gets added to a system-wide
list of explicit "default applications" (mimeapps.list) that will be
checked before falling back to mimeinfo.cache files, but after any
user-specified configuration that might have expliticly set by the user.

[1] https://github.com/Microsoft/vscode/issues/15741

https://phabricator.endlessm.com/T19888