Add X-Eos-Appid to control files
During the transition to Nautilus 3.26.3, we accidentaly lost
the integration with the custom Endless content, due to the
desktop file renaming.

To fix that, add a X-Eos-Appid field to the Debian control
files. This will provide the necessary information to
dh_eoscontent to be able to correctly override the icon,
desktop file and contents of the Nautilus package.

https://phabricator.endlessm.com/T21964